+++ dependency/core/lib/algorithms/schedule_scenario_retrieval/simple_scenario_retrieval.py
@@ -9,7 +9,7 @@ __all__ = ('SimpleScenarioRetrieval',)
 @ClassFactory.register(ClassType.SCH_SCENARIO_RETRIEVAL, alias='simple')
 class SimpleScenarioRetrieval(BaseScenarioRetrieval, abc.ABC):
     def __call__(self, task):
-        scenario = task.get_scenario_data()
+        scenario = task.get_first_scenario_data()
         delay = task.calculate_total_time()
         meta_data = task.get_metadata()
         scenario['delay'] = delay / meta_data['buffer_size']