+++ dependency/core/processor/detector_processor.py
@@ -30,10 +30,11 @@ class DetectorProcessor(Processor):
             success, frame = cap.read()
 
         if len(image_list) == 0:
-            LOGGER.critical('ERROR: image list length is 0')
-            LOGGER.critical(f'Source: {task.get_source_id()}, Task: {task.get_task_id()}')
-            LOGGER.critical(f'file_path: {Context.get_temporary_file_path(task.get_file_path())}')
+            LOGGER.warning(f'[Image list length is 0] Source: {task.get_source_id()} '
+                            f'Task: {task.get_task_id()} '
+                            f'file_path: {Context.get_temporary_file_path(task.get_file_path())}')
             return None
+
         result = self.infer(image_list)
         task = self.get_scenario(result, task)
         task.set_current_content(convert_ndarray_to_list(result))