+++ dependency/core/processor/detector_tracker_processor.py
@@ -31,10 +31,11 @@ class DetectorTrackerProcessor(Processor):
             success, frame = cap.read()
 
         if len(image_list) == 0:
-            LOGGER.critical('ERROR: image list length is 0')
-            LOGGER.critical(f'Source: {task.get_source_id()}, Task: {task.get_task_id()}')
-            LOGGER.critical(f'file_path: {Context.get_temporary_file_path(task.get_file_path())}')
+            LOGGER.warning(f'[Image list length is 0] Source: {task.get_source_id()} '
+                            f'Task: {task.get_task_id()} '
+                            f'file_path: {Context.get_temporary_file_path(task.get_file_path())}')
             return None
+
         result = self.infer(image_list)
         task = self.get_scenario(result, task)
         task.set_current_content(convert_ndarray_to_list(result))
@@ -52,7 +53,8 @@ class DetectorTrackerProcessor(Processor):
             detection_output = self.detector(detection_list)
         result_bbox, result_prob, result_class, result_roi_id = detection_output[0]
         with Timer(f'Tracking / {len(tracking_list)} frame'):
-            tracking_output = self.tracker(tracking_list, detection_list[0], (result_bbox, result_prob, result_class, result_roi_id))
+            tracking_output = self.tracker(tracking_list, detection_list[0],
+                                           (result_bbox, result_prob, result_class, result_roi_id))
         process_output = tracking_output
 
         return process_output